+++ equalsverifier-core/src/main/java/nl/jqno/equalsverifier/internal/reflection/PackageScanner.java
@@ -9,6 +9,8 @@ import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import nl.jqno.equalsverifier.internal.exceptions.ReflectionException;
+
 /** Scans a package for classes. */
 public final class PackageScanner {
 
@@ -45,7 +47,11 @@ public final class PackageScanner {
     }
 
     private static String getResourcePath(URL r) {
-        return rethrow(() -> r.toURI().getPath(), e -> "Could not resolve resource path: " + e.getMessage());
+        String result = rethrow(() -> r.toURI().getPath(), e -> "Could not resolve resource path: " + e.getMessage());
+        if (result == null) {
+            throw new ReflectionException("Could not resolve third-party resource " + r);
+        }
+        return result;
     }
 
     private static List<Class<?>> getClassesInDir(